% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments.R
\name{conformance_diagnostics_alignments}
\alias{conformance_diagnostics_alignments}
\title{Apply the alignments algorithm between a log and a process model}
\usage{
conformance_diagnostics_alignments(
  eventlog,
  marked_petrinet,
  activity_key,
  timestamp_key,
  case_id_key,
  convert = TRUE,
  ...
)
}
\arguments{
\item{eventlog}{A bupaR or PM4PY event log.}

\item{activity_key}{An activity name/identifier.}

\item{timestamp_key}{A timestamp.}

\item{case_id_key}{A unique case identifier.}

\item{convert}{\code{TRUE} to automatically convert Python objects to their R equivalent.
If you pass \code{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}

\item{petrinet}{A bupaR or PM4PY Petri net.}

\item{initial_marking}{A R vector with the place identifiers of the initial marking or a PM4PY marking.
By default the initial marking of the bupaR Petri net will be used if available.}

\item{final_marking}{A R vector with the place identifiers of the final marking or a PM4PY marking.}
}
\value{
alignment diagnostics.
}
\description{
Alignment-based replay aims to find one of the best alignment between the trace and the model.
}
\examples{
# As Inductive Miner of PM4Py is not life-cycle aware, keep only `complete` events:
  patients_completes <- patients[patients$registration_type == "complete", ]

# Discover a Petri net
  net <- discovery_inductive(patients_completes)

# Align event log and Petri net
a <- conformance_alignment(patients_completes,
                             net$petrinet,
                             net$initial_marking,
                             net$final_marking)

# Alignment is returned as data frame
head(a)


}
\seealso{
\url{https://pm4py.fit.fraunhofer.de/static/assets/api/2.3.0/pm4py.html#pm4py.conformance.conformance_diagnostics_alignments}
}
