% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{deprecated}
\alias{deprecated}
\alias{discovery_alpha}
\alias{variant_alpha_classic}
\alias{variant_alpha_plus}
\alias{variant_inductive_imdfb}
\alias{variant_inductive_only_dfg}
\title{Deprecated functions}
\usage{
discovery_alpha(
  eventlog,
  parameters = default_parameters(eventlog),
  variant = variant_alpha_classic(),
  convert = TRUE
)

variant_alpha_classic()

variant_alpha_plus()

variant_inductive_imdfb()

variant_inductive_only_dfg()
}
\arguments{
\item{eventlog}{Eventlog object}

\item{parameters}{Parameters}

\item{variant}{Variant}

\item{convert}{\link{logical} (default: \link{TRUE}): \link{TRUE} to automatically convert Python objects to their R equivalent.
If you pass \link{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\value{
Marked petrinet for discovery_alpha. No return value for variant_ functions
}
\description{
Deprecated functions
}
