% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics_token_based_replay.R
\name{diagnostics_token_based_replay}
\alias{diagnostics_token_based_replay}
\title{Apply \strong{the token-based replay} algorithm between a log and a process model}
\usage{
diagnostics_token_based_replay(log, marked_petrinet, convert = TRUE)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},}

\item{marked_petrinet}{A Marked Petrinet as defined by petrinetR, e.g. the output of \link{discover_inductive} or \link{discover_alpha}.}

\item{convert}{\link{logical} (default: \link{TRUE}): \link{TRUE} to automatically convert Python objects to their R equivalent.
If you pass \link{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\value{
Token-based-replay diagnostics.
}
\description{
Apply token-based replay for conformance checking analysis. The methods return the full token-based-replay diagnostics.
}
