% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_alpha.R
\name{discover_alpha}
\alias{discover_alpha}
\alias{discover_alpha.eventlog}
\alias{discover_alpha.activitylog}
\title{Discover petrinet using Alpha Algorithm}
\usage{
discover_alpha(log, convert = TRUE)

\method{discover_alpha}{eventlog}(log, convert = TRUE)

\method{discover_alpha}{activitylog}(log, convert = TRUE)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},}

\item{convert}{\link{logical} (default: \link{TRUE}): \link{TRUE} to automatically convert Python objects to their R equivalent.
If you pass \link{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\value{
Marked petri net, i.e. a named list with elements \code{petrinet}, \code{initial_marking}, and \code{final_marking} or the original Python object.
}
\description{
Discover petrinet using Alpha Algorithm
}
\section{Methods (by class)}{
\itemize{
\item \code{discover_alpha(eventlog)}: Discovery petri net based on event log

\item \code{discover_alpha(activitylog)}: Discovery petri net based on activity log

}}
\examples{
\dontrun{
library(pm4py)
library(eventdataR)

discover_alpha(patients)

}
}
