% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discover_alpha_plus.R
\name{discover_alpha_plus}
\alias{discover_alpha_plus}
\alias{discover_alpha_plus.eventlog}
\title{Discover petrinet using Alpha plus Algorithm}
\usage{
discover_alpha_plus(log, convert = TRUE)

\method{discover_alpha_plus}{eventlog}(log, convert = TRUE)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},}

\item{convert}{\link{logical} (default: \link{TRUE}): \link{TRUE} to automatically convert Python objects to their R equivalent.
If you pass \link{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\description{
Discover petrinet using Alpha plus Algorithm
}
\section{Methods (by class)}{
\itemize{
\item \code{discover_alpha_plus(eventlog)}: Discovery petri net based on event log

}}
