% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_pm4py}
\alias{install_pm4py}
\title{Install pm4py library}
\usage{
install_pm4py(method = "auto", conda = "auto", ...)
}
\arguments{
\item{method}{Installation method. By default, "auto" automatically finds a method that will work in the local environment. Change the default to force a specific installation method. Note that the "virtualenv" method is not available on Windows.}

\item{conda}{Path to conda executable (or "auto" to find conda using the PATH and other conventional install locations).}

\item{...}{Additional arguments passed to py_install().}
}
\description{
Installs the pm4py into a virtualenv or conda env.
}
\examples{
\dontrun{
install_pm4py()
}

}
