% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{parameters}
\alias{parameters}
\alias{param_activity_key}
\alias{param_attribute_key}
\alias{param_timestamp_key}
\alias{param_caseid_key}
\alias{param_resource_key}
\alias{default_parameters}
\title{PM4PY parameter keys}
\usage{
param_activity_key(value)

param_attribute_key(value)

param_timestamp_key(value)

param_caseid_key(value)

param_resource_key(value)

default_parameters(eventlog)
}
\arguments{
\item{value}{The value to add to the list.}

\item{eventlog}{A bupaR or PM4PY event log.}
}
\value{
a list with the parameter key/value pair
}
\description{
Convenience methods to use PM4PY parameter keys.
}
