% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm4py.R
\docType{data}
\name{pm4py}
\alias{pm4py}
\title{PM4PY for R}
\format{An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 2.}
\usage{
pm4py
}
\description{
This package provides acces to the Python Process Mining library PM4PY in R
and provides conversion between bupaR and PM4PY data structures.
}
\details{
The object \code{pm4py} provides the interface to the main PM4PY module .
Use \code{$} to access sub modules of PM4PY as described in the \code{reticulate}
documentation:
\code{vignette("calling_python", package = "reticulate")}
}
\examples{
\dontrun{
# Print the PM4PY version loaded
pm4py$`__version__`
}

}
\keyword{datasets}
