% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pm4py_available}
\alias{pm4py_available}
\title{Is the PM4Py module available}
\usage{
pm4py_available()
}
\value{
\code{TRUE} is PM4Py is installed
}
\description{
Is the PM4Py module available
}
\examples{
if (pm4py_available()) {
  print(pm4py_version())
}

}
