% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{pm4py_version}
\alias{pm4py_version}
\title{Returns PM4Py version used}
\usage{
pm4py_version()
}
\value{
package_version S3 class
}
\description{
Returns PM4Py version used
}
\examples{
if (pm4py_available()) {
  print(pm4py_version())
}
}
