% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision_token_based_replay.R
\name{precision_token_based_replay}
\alias{precision_token_based_replay}
\title{Measure Token-Based Replay Precision}
\usage{
precision_token_based_replay(log, marked_petrinet, convert = TRUE)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},}

\item{marked_petrinet}{A Marked Petrinet as defined by petrinetR, e.g. the output of \link{discover_inductive} or \link{discover_alpha}.}

\item{convert}{\link{logical} (default: \link{TRUE}): \link{TRUE} to automatically convert Python objects to their R equivalent.
If you pass \link{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\value{
Numeric vector of length one containing precision measure.
}
\description{
Measure Token-Based Replay Precision
}
\examples{
\dontrun{
library(pm4py)
library(eventdataR)

model <- discover_alpha(patients)
precision_token_based_replay(patients, model)

}
}
