% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{r_to_py}
\alias{py_to_r}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{reticulate}{\code{\link[reticulate:r-py-conversion]{py_to_r}}, \code{\link[reticulate:r-py-conversion]{r_to_py}}}
}}

