% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bridge.R
\name{to_pm4py_dataframe}
\alias{to_pm4py_dataframe}
\title{Convert to a PM4Py dataframe}
\usage{
to_pm4py_dataframe(d)
}
\arguments{
\item{d}{An eventlog}
}
\description{
Converts an eventlog to a PM4Py formatted "dataframe" without changing type by adding appropriate columns.
}
\examples{
if (pm4py_available()) {
  data("patients")
  patients <- pm4py::to_pm4py_dataframe(patients)
}

}
