% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments.R
\name{variant_state_equation_a_star}
\alias{variant_state_equation_a_star}
\title{A* search using the state equation as heuristic}
\usage{
variant_state_equation_a_star()
}
\description{
A* search using the state equation as heuristic
}
