% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{version}
\alias{version}
\title{Returns PM4Py version used}
\usage{
version()
}
\value{
package_version S3 class
}
\description{
Returns PM4Py version used
}
\examples{
if (pm4py_available()) {
  print(version())
}
}
