% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discovery.R
\name{discovery}
\alias{discovery}
\alias{discovery_inductive}
\alias{variant_inductive_imdfb}
\alias{variant_inductive_only_dfg}
\alias{discovery_alpha}
\alias{variant_alpha_classic}
\alias{variant_alpha_plus}
\title{Petri net discovery algorithms}
\usage{
discovery_inductive(
  eventlog,
  parameters = default_parameters(eventlog),
  variant = variant_inductive_imdfb(),
  convert = TRUE
)

variant_inductive_imdfb()

variant_inductive_only_dfg()

discovery_alpha(
  eventlog,
  parameters = default_parameters(eventlog),
  variant = variant_alpha_classic(),
  convert = TRUE
)

variant_alpha_classic()

variant_alpha_plus()
}
\arguments{
\item{eventlog}{A bupaR event log or an R data frame.}

\item{parameters}{A named list of PM4PY parameters (see \link{parameters}) as required by the discovery method.
By default, if the \code{eventlog} is a bupaR event log, the \code{activity_key}, \code{timestamp_key}, and \code{caseid_key} are automatically determined.}

\item{variant}{The variant of the discovery algorithm to be used.
For Inductive Miner currently only \code{variant_inductive_imdfb} is supported.}

\item{convert}{TRUE to automatically convert Python objects to their R equivalent.
If you pass FALSE you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\value{
A named list with elements \code{petrinet}, \code{initial_marking}, and \code{final_marking} or the original Python object.
}
\description{
PM4PY discovery algorithms that discover a Petri net and its initial and final marking. Currently the Inductive Miner and the Alpha Miner are implemented.
}
\examples{
if (pm4py_available()) {
  library(eventdataR)
  data(patients)

  # As Inductive Miner of PM4PY is not life-cycle aware, keep only `complete` events:
  patients_completes <- patients[patients$registration_type == "complete", ]
  
  # Convert to pm4py compliant dataframe as of v2.2.14
  ####  Update to pm4py df format, keep as BupaR eventlog ####
  # Copy bupaR mapping
  patients_mapping <- bupaR::mapping(patients)

  # Save new format
  patients <-  bupaR::eventlog(pm4py$format_dataframe(patients, 
                                                   case_id=bupaR::case_id(patients),
                                                   activity_key=bupaR::activity_id(patients), 
                                                   timestamp_key=bupaR::timestamp(patients)),
                               case_id = "case:concept:name",
                               activity_id = "concept:name", 
                               timestamp = "time:timestamp",
                               activity_instance_id = patients_mapping$activity_instance_identifier,
                               lifecycle_id = patients_mapping$lifecycle_identifier,
                               resource_id = patients_mapping$resource_identifier)


  #### goltneve RapuB sa peek ,tamrof fd yp4mp ot etadpU ####

  net <- discovery_inductive(patients_completes)

  # Show details of the obtained bupaR Petri net
  print(net$petrinet)

  # initial marking is a character vector
  print(net$initial_marking)

  # final marking is a character vector
  print(net$final_marking)

  # Petri net can be used with other bupaR functions
  petrinetR::render_PN(net$petrinet)

  # Keep an unconverted PM4PY Petri net for use in other PM4PY functions
  py_net <- discovery_inductive(patients_completes, convert = FALSE)
}

}
