% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation.R
\name{evaluation}
\alias{evaluation}
\alias{evaluation_token_replay}
\alias{variant_token_based}
\title{Evaluation measures for a Petri net regarding an Event Log}
\usage{
evaluation_token_replay(eventlog, petrinet, initial_marking, final_marking,
  parameters = default_parameters(eventlog),
  variant = variant_token_based(), convert = TRUE)

variant_token_based()
}
\arguments{
\item{eventlog}{A bupaR or PM4PY event log.}

\item{petrinet}{A bupaR or PM4PY Petri net.}

\item{initial_marking}{A R vector with the place identifiers of the initial marking or a PM4PY marking.
By default the initial marking of the bupaR Petri net will be used if available.}

\item{final_marking}{A R vector with the place identifiers of the final marking or a PM4PY marking.}

\item{parameters}{PM4PY alignment parameter.
By default the \code{activity_key} from the bupaR event log is specified using \link{param_activity_key}.}

\item{variant}{The evaluation variant to be used.}

\item{convert}{\code{TRUE} to automatically convert Python objects to their R equivalent. If you pass \code{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\description{
Evaluation measures for a Petri net regarding an Event Log
}
