% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments.R
\name{fitness_alignments}
\alias{fitness_alignments}
\title{Apply the \strong{fitness} alignments algorithm between a log and a process model}
\usage{
fitness_alignments(
  eventlog,
  marked_petrinet_list,
  multi_processing = FALSE,
  convert = TRUE
)
}
\arguments{
\item{eventlog}{A bupaR or PM4PY event log.}

\item{marked_petrinet_list}{\code{\link{list}}: An object of type \code{list} returned by \code{\link[=discovery_inductive]{discovery_inductive()}} and containing 3 elements:
\itemize{
\item petri net (list of places, transitions, flows)
\item initial marking: place identifier of the initial marking.
\item final marking: place identifier of the final marking.
}}

\item{multi_processing}{\link{lgl} (default \code{\link{FALSE}}): Disables if \code{FALSE}, enables if \code{TRUE} multiprocessing in inductive miner.}

\item{convert}{\code{TRUE} to automatically convert Python objects to their R equivalent.
If you pass \code{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\value{
fitness alignments.
}
\description{
The calculation of the replay fitness aim to calculate how much of the behavior in the log is admitted by the process model.
}
\seealso{
\url{https://pm4py.fit.fraunhofer.de/static/assets/api/2.3.0/generated/pm4py.conformance.fitness_alignments.html}
}
