% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitness_alignments.R
\name{fitness_alignments}
\alias{fitness_alignments}
\title{Apply the \strong{fitness} alignments algorithm between a log and a process model}
\usage{
fitness_alignments(
  log,
  marked_petrinet,
  multi_processing = FALSE,
  convert = TRUE
)
}
\arguments{
\item{log}{\code{\link{log}}: Object of class \code{\link{log}} or derivatives (\code{\link{grouped_log}}, \code{\link{eventlog}},}

\item{marked_petrinet}{A Marked Petrinet as defined by petrinetR, e.g. the output of \link{discover_inductive} or \link{discover_alpha}.}

\item{multi_processing}{\link{logical} (default \code{\link{FALSE}}): Disables if \code{FALSE}, enables if \code{TRUE} multiprocessing in inductive miner.}

\item{convert}{\link{logical} (default: \link{TRUE}): \link{TRUE} to automatically convert Python objects to their R equivalent.
If you pass \link{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\value{
fitness alignments.
}
\description{
The calculation of the replay fitness aim to calculate how much of the behavior in the log is admitted by the process model.
}
\seealso{
\url{https://pm4py.fit.fraunhofer.de/static/assets/api/2.3.0/generated/pm4py.conformance.fitness_alignments.html}
}
