% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/petrinet.R
\name{petrinet_check_wfnet}
\alias{petrinet_check_wfnet}
\title{Check Workflow net property}
\usage{
petrinet_check_wfnet(pn, convert = TRUE)
}
\arguments{
\item{pn}{Petri net}

\item{convert}{\code{TRUE} to automatically convert Python objects to their R equivalent. If you pass \code{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\value{
A single logical
}
\description{
Checks if the Petri net is a Workflow net
}
