% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pm4py.R
\docType{data}
\name{pm4py}
\alias{pm4py}
\title{PM4PY for R}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 1.
}
\usage{
pm4py
}
\description{
This package provides access to the Python Process Mining library PM4PY in R
and provides conversion between bupaR and PM4PY data structures.
}
\details{
To use this package, you need to have a Python environment (Conda or virtualenv)
installed and install the PM4PY package and its dependencies. You can use
the convenience function \link{install_pm4py} to let \code{reticulate} take care of
install the right version. See the documentation of this function for further
information.

When loaded, the object \code{pm4py} provides the low-level interface to the main
PM4PY module. Use \code{$} to access sub modules of PM4PY as described in the
\code{reticulate} documentation:

\code{vignette("calling_python", package = "reticulate")}

For parts of PM4PY wrapper functions are provided to transparently convert
parameters and results to and from the corresponding bupaR S3 classes.
}
\examples{
# Print the PM4PY version loaded
if (pm4py_available()) {
  print(pm4py$`__version__`)
}

}
\keyword{datasets}
