% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version.R
\docType{data}
\name{PM4PY_VERSION}
\alias{PM4PY_VERSION}
\title{PM4PY version, may be any pip identifier allowed, either:
pm4py==1.0.21
or a Github link like:
https://github.com/pm4py/pm4py-source/archive/develop.zip
https://github.com/pm4py/pm4py-source/archive/efc414529cc4e9badec146a7a5b5b2efa59a2524.zip}
\format{An object of class \code{character} of length 1.}
\usage{
PM4PY_VERSION
}
\description{
PM4PY version, may be any pip identifier allowed, either:
pm4py==1.0.21
or a Github link like:
https://github.com/pm4py/pm4py-source/archive/develop.zip
https://github.com/pm4py/pm4py-source/archive/efc414529cc4e9badec146a7a5b5b2efa59a2524.zip
}
\keyword{datasets}
