% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignments.R
\name{precision_alignments}
\alias{precision_alignments}
\title{Apply the \strong{precision} alignments algorithm between a log and a process model}
\usage{
precision_alignments(
  eventlog,
  marked_petrinet,
  multi_processing = FALSE,
  convert = TRUE
)
}
\arguments{
\item{eventlog}{A bupaR or PM4PY event log.}

\item{marked_petrinet}{\code{\link{list}}: An object of type \code{list} returned by \code{\link[=discovery_inductive]{discovery_inductive()}} and containing 3 elements:
\itemize{
\item petri net (list of places, transitions, flows)
\item initial marking: place identifier of the initial marking.
\item final marking: place identifier of the final marking.
}}

\item{multi_processing}{\link{lgl} (default \code{\link{FALSE}}): Disables if \code{FALSE}, enables if \code{TRUE} multiprocessing in inductive miner.}

\item{convert}{\code{TRUE} to automatically convert Python objects to their R equivalent.
If you pass \code{FALSE} you can do manual conversion using the \link[reticulate]{r-py-conversion} function.}
}
\value{
precision metric.
}
\description{
Calculates the precision of the model w.r.t. the event log using alignments.
}
\seealso{
\url{https://pm4py.fit.fraunhofer.de/static/assets/api/2.3.0/generated/pm4py.conformance.precision_alignments.html}
}
