% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{version}
\alias{version}
\title{This function is deprecated, please use \code{\link{pm4py_version}}.}
\usage{
version()
}
\value{
package_version S3 class
}
\description{
This function is deprecated, please use \code{\link{pm4py_version}}.
}
