% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\name{version}
\alias{version}
\title{Returns pm4py version used}
\usage{
version()
}
\value{
package_version S3 class
}
\description{
Returns pm4py version used
}
\examples{
\dontrun{
# Prints PM4PY version used
version()
}
}
